switch kwargs to arguments in images_build.py
This commit converts the build function in images_build.py to
regular typed function arguments. This requires changes also
to related tests because of parameters that now have type
pathlib.Path instead of string, and this means that this
change will break other's code.
Note: build will close fileobj (when possible), if it is an
io.BytesIO object passed from a caller function then it needs
to be copied, otherwise the function will close the parent's object.

Signed-off-by: Federico Rizzo <fulminemizzega@yahoo.it>